requirements.txt
# HTTP client
requests>=2.31.0

# LLM API (OpenAI-compatible interface used by Anthropic, OpenAI, and others)
anthropic>=0.25.0
openai>=1.0.0

# REST API interface
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# Data validation
pydantic>=2.0.0

# Environment variable management
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-cov>=5.0.0
responses>=0.25.0
